requirements.txt
fastapi==0.111.0
nltk==3.8.1
pandas==2.2.2
Pillow==9.3.0
Pillow==10.3.0
pydantic==2.7.1
scikit_learn==1.5.0
torch==2.4.0.dev20240521+cu124
torchvision==0.19.0.dev20240521+cu124
